Created pushcontainer-check to see if the current container needs to be built.
Since changes to some code could not lead to changes in container
code (ex. K8s files for example), then the PVER could increment (i.e a
git commit) but no container built and labeled in the remote repo. So,
this checks for remote container existence.

NOTE: this ONLY works for AWS repos... bummer.
NOTE2: don't really need to rebuild the container, which this will do,
  since the code didn't change. Docker will see no difference anyway and
  just apply a different label. In the future, could make two calls
  before building to prevent an unnecessary build.