Lots of changes.
The bigged change is stage modifiers. Most everything else is related to
this in one way or another (or just a bug fix). I think I call them
devmodes in the code but this will be changed. Checking in the code now
as I am going to fix this code to make it more granular.

Major:
* introduced stage modifiers. the allow local development of a
microservice with the effect of being in the cluster. It port forwards
necessary ports into the cluster (i.e. other microservices or
databases). With this, a microservice can pull information from a
running system and you can overserve how it reacts.
  * The magic of getting this to work is how it overwrites configuration
  information in the build stage. So. a configuration of "plugin.STAGE.url" would point
  to something in the cluster. When switching to local, the wov-build
  script finds the modifiers for "plugin.STAGE-local.url" and overwrites
  the value. This adds some non-insignificant additional overhead in time. It is also nice because I can write configuration information for local operation right next to the other configuration information in the file.
  * However, the limitation is you can't port forward from
other microservices to this microservice (lots of discussion about this
online because why the fudge didn't they allow this? (I know...
security).
  * not granular enough though, since I need to work local on one microservice but remote on others. this manages the port forwarding through scripts for now.
  * this should make testing MUCH easier to accomplish however in a more
scriptable form.
  * the other problem is changing modes causes git changes, causing a
  new push and a new deploy... very annoying. So, introducing a new
  config_local.json file next, where local changes are placed that are
  not tracked by git. This will undoubtedly be used by other scripts in
  time.
  * created a process to manage portforwards on the localhost... needed
  for when you toggle devmodes.

Other Changes:
* fixed some issues with scripts rewriting global vars when other
scripts are run in hte same namespace
* ignore namespace added to some scripts (helpful for working between
    namespaces)
* bug fix for push scripts that broke pushes when AWS keys were outdated
* added mongodb to the wov-db-connect script... so to my scripts,
  postgres and mongodb look the same
* Can deploy only certain microservices now, instead of multiple
* Huge bug in ConfigMap/Secret creation that I have no idea how it was
working beforehand. Bash crap and quoting was to blame.
* wov-ed can open a editor in new window now with -x switch
* added more attributes to deployments so they are clearer what version
the deployment is. wov-ns now shows lables.
* GREATLY sped up wov-push. It was taking, on a very good internet
connection, an additional 30 seconds (several minutes at Panera). Now,
it gets it down to about 20 seconds total, 10 seconds being network
delays. much less disruptive to development.