Built the stage modifiers in and devmodes out.
So, this generates
conf-${STAGEMODIFIER} and secret/current-${STAGEMODIFIER}.json files in
parallel, when wov-build --mod X is used. Then, the wov-env --mod X is
used to select which modifier to use (off by default). This allows local
scripts to be invoked with local properties but everything to the
cluster is exactly the same.

The wov-portforward command is used to forward to services in a cluster,
so we can forward to cluster databases or a particular pod.

Together, we now have granularity enough to run certain services in the
cluster and others locally. i.e. plemwl is in the cluster, listening to
the API and webserver, while plemrl is local but responding to plemwl
and using the cluster's database.