Major changes to how this pushes deployments with the new configuration.
The major change is pushing the exact ConfigMap and Secrets needed for a
pod. This also generates those for the 'vh' pod as well. It knows what
to make based upon the container recipes and it gets the variables from
the Config class in Nodejs or a script file. This greatly clears up a
lot of headaches i've had with configuration.