Too many changes. Need to completely change how WovTools handles configuration mods for local and clusters/stages.
See github issue:31 - https://github.com/woveon/wovtools/issues/31

Basically, these are longstanding issues all tied together, and
I finally have a plan. This came to a head since I can't push until I
fix this.

So, here is a full checkin that hasn't been done in a while, with a lot
of junk, so that I can checkpoint this code as I move forward.

...wish me luck.